backend/requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.30.6
pydantic-settings==2.6.1
requests==2.32.3
langdetect==1.0.9
qdrant-client==1.9.2
sentence-transformers==3.0.1
pypdf==4.3.1
python-docx==1.1.2
rank-bm25==0.2.2
numpy==1.26.4
FlagEmbedding==1.2.10
scikit-learn==1.4.2
prometheus-fastapi-instrumentator
